feat: 通用逻辑迁移utils
