+++ src/utils/format/format-encipher.ts
@@ -1,0 +1,3 @@
+// 加密解密工具
+
+export function formatEncipherTest() {}
