+++ vite.config.ts
@@ -70,10 +70,7 @@ export default ({ mode, command }) => {
             './src/views/overview/Overview.vue',
             './src/views/overview/OverviewSearch.vue',
             './src/views/overview/OverviewSub.vue',
-            './src/views/tool/FormatHump.vue',
-            './src/views/tool/FormatUnderline.vue',
-            './src/views/tool/FormatLowerCase.vue',
-            './src/views/tool/FormatUpperCase.vue',
+            //
             './src/views/tool/FormatTime.vue',
             './src/views/tool/FormatTimeUnix.vue',
             './src/views/tool/FormatTextJson.vue'
